+++ simapp/v2/simdv2/cmd/commands.go
@@ -22,7 +22,6 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/debug"
-	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/keys"
 	"github.com/cosmos/cosmos-sdk/client/rpc"
 	"github.com/cosmos/cosmos-sdk/server"
@@ -49,12 +48,8 @@ func (t *temporaryTxDecoder) DecodeJSON(bz []byte) (transaction.Tx, error) {
 	return t.txConfig.TxJSONDecoder()(bz)
 }
 
-func newApp(
-	logger log.Logger,
-	viper *viper.Viper,
-) serverv2.AppI[transaction.Tx] {
-	sa := simapp.NewSimApp(logger, viper)
-	return sa
+func newApp(logger log.Logger, viper *viper.Viper) serverv2.AppI[transaction.Tx] {
+	return simapp.NewSimApp(logger, viper)
 }
 
 func initRootCmd(
@@ -184,12 +179,6 @@ func appExport(
 	viper *viper.Viper,
 	modulesToExport []string,
 ) (servertypes.ExportedApp, error) {
-	// this check is necessary as we use the flag in x/upgrade.
-	// we can exit more gracefully by checking the flag here.
-	homePath, ok := viper.Get(flags.FlagHome).(string)
-	if !ok || homePath == "" {
-		return servertypes.ExportedApp{}, errors.New("application home not set")
-	}
 	// overwrite the FlagInvCheckPeriod
 	viper.Set(server.FlagInvCheckPeriod, 1)